+++ peon_engine/Source/Video/OpenGL/OpenGLWindowModule.hpp
@@ -22,7 +22,7 @@ namespace graphics
 class OpenGLWindowModule : public IWindowModule
 {
 public:
-    OpenGLWindowModule(Engine* const engine, const std::string& title, const uint16_t& width, const uint16_t& height, const bool& isFullscreen);
+    OpenGLWindowModule(Engine* const engine, const std::string& title, const uint32_t& width, const uint32_t& height, const bool& isFullscreen);
     ~OpenGLWindowModule();
 
     /*
